Added basic spawning of & communication with stub
Since w3wm uses DLL injection to get keyboard input before other
programs, it needs to address the fact that w3wm cannot be a single
executable on 64-bit machines. Instead, there needs to exist a 32-bit
executable injecting 32-bit code, and a 64-bit executable injecting
64-bit code.

As there is always a need for a 32-bit executable, the main w3wm
executable is 32-bit. On 64-bit machines, this executable will spawn a
64-bit stub to perform the 64-bit injection.

This change takes care of the spawning, killing, and communication with
the 64-bit stub.